Rename objects/common.py to objects/metadata.py
The only classes left in objects/common.py were metadata definition
classes. As such, it made sense to rename it to denote this.
