+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/data_source.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import List, Optional, Sequence
 
 from dbt_semantic_interfaces.objects.base import ModelWithMetadataParsing, HashableBaseModel
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.objects.elements.dimension import Dimension
 from dbt_semantic_interfaces.objects.elements.entity import Entity
 from dbt_semantic_interfaces.objects.elements.measure import Measure
