+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/elements/dimension.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Optional
 
 from dbt_semantic_interfaces.objects.base import HashableBaseModel, ModelWithMetadataParsing
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.references import DimensionReference, TimeDimensionReference
 from dbt_semantic_interfaces.enum_extension import ExtendedEnum
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
