+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/elements/entity.py
@@ -6,7 +6,7 @@ from dbt_semantic_interfaces.objects.base import (
     HashableBaseModel,
     ModelWithMetadataParsing,
 )
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.enum_extension import ExtendedEnum
 from dbt_semantic_interfaces.references import EntityReference
