+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/elements/measure.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Optional, List
 from dbt_semantic_interfaces.objects.aggregation_type import AggregationType
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.objects.base import ModelWithMetadataParsing, HashableBaseModel
 from dbt_semantic_interfaces.references import MeasureReference, TimeDimensionReference
