+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/metric.py
@@ -9,7 +9,7 @@ from dbt_semantic_interfaces.objects.base import (
     PydanticCustomInputParser,
     PydanticParseableValueType,
 )
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.objects.constraints.where import WhereClauseConstraint
 from dbt_semantic_interfaces.references import MeasureReference, MetricReference
 from dbt_semantic_interfaces.enum_extension import ExtendedEnum
