+++ metricflow/model/validations/validator_helpers.py
@@ -12,7 +12,7 @@ import click
 from pydantic import BaseModel, Extra
 
 from dbt_semantic_interfaces.objects.base import FrozenBaseModel
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.references import (
