+++ metricflow/test/model/parsing/test_metadata_parsing.py
@@ -5,7 +5,7 @@ from typing import Sequence
 
 import pytest
 
-from dbt_semantic_interfaces.objects.common import Metadata
+from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.objects.elements.measure import Measure
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.parsing.yaml_loader import YamlConfigLoader
